+++ .gitignore
@@ -60,5 +60,4 @@ target/
 .ipynb_checkpoints
 .DS_Store
 .idea
-venv
 *~
