--- README.md
@@ -1,3 +1,0 @@
-# kaggle-santander
-
-Boulder Data Science at it again with another Kaggle competition!
