+++ environment_osx.yml
@@ -1,0 +1,67 @@
+name: kaggle-santander
+dependencies:
+- appnope=0.1.0=py27_0
+- backports_abc=0.4=py27_0
+- cycler=0.10.0=py27_0
+- decorator=4.0.9=py27_0
+- freetype=2.5.5=0
+- ipykernel=4.3.1=py27_0
+- ipython=4.1.2=py27_0
+- ipython_genutils=0.1.0=py27_0
+- ipywidgets=4.1.1=py27_0
+- jinja2=2.8=py27_0
+- jsonschema=2.4.0=py27_0
+- jupyter=1.0.0=py27_1
+- jupyter_client=4.2.1=py27_0
+- jupyter_console=4.1.1=py27_0
+- jupyter_core=4.1.0=py27_0
+- libpng=1.6.17=0
+- markupsafe=0.23=py27_0
+- matplotlib=1.5.1=np110py27_0
+- mistune=0.7.2=py27_0
+- mkl=11.3.1=0
+- nbconvert=4.1.0=py27_0
+- nbformat=4.0.1=py27_0
+- notebook=4.1.0=py27_1
+- numpy=1.10.4=py27_0
+- openssl=1.0.2g=0
+- pandas=0.18.0=np110py27_0
+- path.py=8.1.2=py27_1
+- pexpect=3.3=py27_0
+- pickleshare=0.5=py27_0
+- pip=8.1.0=py27_0
+- ptyprocess=0.5=py27_0
+- pygments=2.1.1=py27_0
+- pyparsing=2.0.3=py27_0
+- pyqt=4.11.4=py27_1
+- python=2.7.11=0
+- python-dateutil=2.5.0=py27_0
+- python.app=1.2=py27_4
+- pytz=2016.1=py27_0
+- pyzmq=15.2.0=py27_0
+- qt=4.8.7=1
+- qtconsole=4.1.1=py27_1
+- readline=6.2=2
+- scikit-learn=0.17.1=np110py27_0
+- scipy=0.17.0=np110py27_0
+- setuptools=20.2.2=py27_0
+- simplegeneric=0.8.1=py27_0
+- singledispatch=3.4.0.3=py27_0
+- sip=4.16.9=py27_0
+- six=1.10.0=py27_0
+- sqlite=3.9.2=0
+- ssl_match_hostname=3.4.0.2=py27_0
+- terminado=0.5=py27_1
+- tk=8.5.18=0
+- tornado=4.3=py27_0
+- traitlets=4.1.0=py27_0
+- wheel=0.29.0=py27_0
+- zlib=1.2.8=0
+- pip:
+  - backports-abc==0.4
+  - backports.ssl-match-hostname==3.4.0.2
+  - ipython-genutils==0.1.0
+  - jupyter-client==4.2.1
+  - jupyter-console==4.1.1
+  - jupyter-core==4.1.0
+
